(#232) Pagure split and __init__
